+++ js/components/ListHeading/Link.js
@@ -1,16 +1,16 @@
 import styled from 'styled-components';
 import { Link } from 'react-router';
-import constants from '../../constants';
+import { colors } from '../../constants';
 
 export default styled(Link)`
   position: absolute;
   top: 0;
   left: 0;
   padding: 1em;
-  border-right: 1px solid ${constants.lightGrey};
+  border-right: 1px solid ${colors.lightGrey};
 
   &:hover {
-    background-color: ${constants.lightGrey};
+    background-color: ${colors.lightGrey};
   	cursor: pointer;
   }
 `;
