+++ js/components/ListHeading/Wrapper.js
@@ -1,10 +1,10 @@
 import styled from 'styled-components';
-import constants from '../../constants';
+import { colors } from '../../constants';
 
 export default styled.div`
   padding: 1em;
   margin: 0;
-  border-bottom: 1px solid ${constants.lightGrey};
+  border-bottom: 1px solid ${colors.lightGrey};
   text-align: center;
   width: 100%;
   position: relative;
