+++ js/components/PluginListItem/Name.js
@@ -1,9 +1,9 @@
 import styled from 'styled-components';
-import constants from '../../constants';
+import { colors } from '../../constants';
 
 export default styled.h2`
   display: inline-block;
-  color: ${constants.textColor};
+  color: ${colors.textColor};
   margin-top: 0;
   margin-bottom: 0;
   margin-right: 1em;
