+++ js/components/PluginListItem/Wrapper.js
@@ -1,16 +1,16 @@
 import styled from 'styled-components';
-import constants from '../../constants';
+import { colors } from '../../constants';
 
 export default styled.a`
   display: block;
   padding: 0.75em 1.5em;
-  border-bottom: 1px solid ${constants.lightGrey};
+  border-bottom: 1px solid ${colors.lightGrey};
   text-decoration: none;
   &:hover {
-    background-color: ${constants.lightGrey};
+    background-color: ${colors.lightGrey};
   }
   &:focus {
-    background-color: ${constants.lightGrey};
+    background-color: ${colors.lightGrey};
   }
   &:last-child {
     border-radius: 0 0 5px 5px;
