+++ js/components/SearchField/Link.js
@@ -2,7 +2,7 @@ import React from 'react';
 import Link from 'react-router/lib/Link';
 import styled from 'styled-components';
 
-import constants from '../../constants';
+import { colors } from '../../constants';
 
 const StyledLink = styled(Link)`
   position: absolute;
@@ -10,13 +10,13 @@ const StyledLink = styled(Link)`
   right: 0;
   transform: translateY(-50%);
   font-size: 1em;
-  color: ${constants.textColor};
+  color: ${colors.textColor};
   text-decoration: none;
   padding: 1em;
-  border-left: 1px solid ${constants.lightGrey};
+  border-left: 1px solid ${colors.lightGrey};
 
   &:hover {
-    background-color: ${constants.lightGrey};
+    background-color: ${colors.lightGrey};
   }
 `;
