+++ js/components/SearchField/Wrapper.js
@@ -1,9 +1,9 @@
 import styled from 'styled-components';
-import constants from '../../constants';
+import { colors } from '../../constants';
 
 export default styled.div`
   width: 100%;
-  border-bottom: 1px solid ${constants.lightGrey};
+  border-bottom: 1px solid ${colors.lightGrey};
   position: relative;
   padding-right: 2em;
 `;
