+++ js/components/Spinner/Wrapper.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export default styled.div`
+  margin: 2em auto;
+  width: 40px;
+  height: 40px;
+  position: relative;
+`;
